+++ src/components/Header/RightMenu/AuthLink.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Link as MaterialLink } from "@mui/material";
+
+const AuthLink = () => {
+    return (
+        <>
+            <MaterialLink className="link">Account</MaterialLink>
+        </>
+    );
+};
+
+export default AuthLink;
